Null checks format changed across all files. ObservableMap - improvements and correctness changes. OutlinedLabel - no longer has a fontsize, refactoring in the same class.
IntermissionTimer no longer utilises the UI thread, it's now using the ThreadPool, to avoid cluttering the UI thread and improve performance.

Subtitles Management -
fixed an unintended feature where a paused media would resume when  the slider is dragged while the MediaElement would be paused, from now on this is being synchronized with the MediaElement state to allow more flexibility.
Intoduced new method which forces the next tick of the internal timer to wait in case it's 10>ms early, it's also being synchronized with the MediaElement.Position.
Minor changes to the other subtitle related classes.